* moving raw blob stuff from gui.cpp to different file * BasicStuff now does not hold a pointer anymore, but a solid object (saving 8 bytes per frame doing this, and it is semantically sound)
