+++ Application/src/tracker/python/GPURecognition.cpp
@@ -35,7 +35,7 @@ namespace pybind11 {
             PYBIND11_TYPE_CASTER(cmn::Image::Ptr, _("Image::Ptr"));
 
             // Conversion part 1 (Python -> C++)
-            bool load(py::handle src, bool convert)
+            bool load(py::handle , bool )
             {
                 /*if ( !convert and !py::array_t<T>::check_(src) )
                   return false;
@@ -59,7 +59,7 @@ namespace pybind11 {
             }
 
             //Conversion part 2 (C++ -> Python)
-            static py::handle cast(const cmn::Image::Ptr& src, py::return_value_policy policy, py::handle parent)
+            static py::handle cast(const cmn::Image::Ptr& src, py::return_value_policy, py::handle)
             {
 
                 std::vector<size_t> shape{ src->rows, src->cols, src->dims };
@@ -85,7 +85,7 @@ namespace pybind11 {
                 return a.release();
             }
             
-            static py::handle cast(const cmn::Image::UPtr& src, py::return_value_policy policy, py::handle parent)
+            static py::handle cast(const cmn::Image::UPtr& src, py::return_value_policy , py::handle )
             {
 
                 std::vector<size_t> shape{ src->rows, src->cols, src->dims };
@@ -107,13 +107,13 @@ namespace pybind11 {
         PYBIND11_TYPE_CASTER(cmn::Image::UPtr, _("Image::UPtr"));
 
         // Conversion part 1 (Python -> C++)
-        bool load(py::handle src, bool convert)
+        bool load(py::handle, bool)
         {
             return false;
         }
 
         //Conversion part 2 (C++ -> Python)
-        static py::handle cast(const cmn::Image::UPtr& src, py::return_value_policy policy, py::handle parent)
+        static py::handle cast(const cmn::Image::UPtr& src, py::return_value_policy, py::handle)
         {
 
             std::vector<size_t> shape{ src->rows, src->cols, src->dims };
